Now displays time on x-axis in "m:s" format as well as being able to see either iat dev or bitrate on graph.

+++ Middleware/analyzerpcapmiddleware.h
@@ -41,7 +41,7 @@ public:
 
 signals:
     void status(AnalyzerStatus status);
-    void bitrateStatus(qint64 bitrate, qint64 duration);
+    void bitrateStatus(double bitrate, qint64 duration);
 
 protected slots:
     void run();
